+++ ZFSin/spl/module/spl/spl-vmem.c
@@ -432,7 +432,7 @@ uint64_t spl_bucket_tunable_small_span = 0;
 // for XAT & XATB visibility into VBA queue
 static _Atomic uint32_t spl_vba_threads[VMEM_BUCKETS] = { 0 };
 static uint32_t vmem_bucket_id_to_bucket_number[NUMBER_OF_ARENAS_IN_VMEM_INIT] = { 0 };
-boolean_t spl_arc_no_grow(uint32_t, boolean_t, kmem_cache_t **);
+boolean_t spl_arc_no_grow(size_t, boolean_t, kmem_cache_t **);
 _Atomic uint64_t spl_arc_no_grow_bits = 0;
 uint64_t spl_arc_no_grow_count = 0;
 
@@ -3745,7 +3745,7 @@ vmem_bucket_number_arc_no_grow(const uint32_t size)
 }
 
 boolean_t
-spl_arc_no_grow(uint32_t size, boolean_t buf_is_metadata, kmem_cache_t **zp)
+spl_arc_no_grow(size_t size, boolean_t buf_is_metadata, kmem_cache_t **zp)
 {
 	const uint16_t b = vmem_bucket_number_arc_no_grow(size);
